Seaborn.py
Create a boxplot using the boxplot function, where the x-axis represents different days of the week and the y-axis represents the total bill. The boxplot provides a visual summary of the minimum, first quartile, median, third quartile and maximum of the total bill for each day.

We then set a title for the plot and finally, display it using plt.show()'.